Add urlconf reload to SeleniumLiveTestCase.
This commit modifies the setUpClass method of SeleniumLiveTestCase
to look for a test_templates list of path/template tuples in the
derived class that allows an arbitrary list of pages to be added
for each set of unit tests.  Modifying the urlconf must happen before
the live server thread is forked, so this functionality now exists
in the class constructor.