Move render_entry_log into try block.
The wait_for decorator is applied to test methods for cases where a
page may not be fully loaded. The intent is to catch any exception
that results from a page element that cannot be found and retry (for
some time) to see if the unfound element loads eventually.

Meanwhile, the render_entry_log() method call was added to wait_for to
log the state of the page (in text and optionally as an image) at the
beginning of any wrapped call.

But this led to a bug in that render_entry_log() itself can fail if
the page is not yet fully loaded, (no "//body" element can be
found). To fix this bug, in this commit the render_entry_log() call is
moved into the 'try' block so that that exception will also be caught
and retried.