Add should_be_logged_in_as to AuthenticationTestMixin.
The login method was failing intermittently due to the fact that
there was no test to see if the login was complete.  This led to
race conditions specifically when self.get_page() was called.

This commit adds a new should_be_logged_in_as method for external
users and calls it as the last step of self.login().  This code
assumes that database sessions are enabled and that if a session
with that userid exists.