Add method to NavigationTestMixin to hover over a button or link.
The hover_over_button method is added to perform hover actions
in the browser. These actions are needed to test UI elements such
as dropdown menus, which only appear when hovering over a button
or link.

This method was added after observing the
test_switching_project_views_with_project_dropdown test was failing
on Firefox. That test initially tried to click on the dropdown
link to render the menu items. That approach worked when testing
on Chrome, but failed on Firefox.

Using the hover_over_button method should work for both browsers
and performs the correct actions needed for that test.